requirements.txt
absl-py==1.0.0
astunparse==1.6.3
blinker==1.6.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.1.3
colorama==0.4.6
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.3.1
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.5.1
nltk==3.8.1
numpy==1.22.2
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
regex==2023.3.23
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.3.1
wrapt==1.13.3
